requirements.txt
streamlit
pandas
load_dotenv
sqlalchemy
psycopg2
os
time
seaborn
matplotlib
plotly
nltk
string
re
selenium
bs4
numpy
requests
warnings
datetime
urllib